revert enum move
